requests: break out Requests interface and classes
